search: move query visitor to own file
